korrekturä
Ha z maindeck iz gamemasterscript inägmergt u ir main menue scene öpis glöscht
