--- Assets/Scripts/ActiveDeck.cs
@@ -1,20 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ActiveDeck : MonoBehaviour
-{
-    [SerializeField] public List<GameObject> content;
-    //An Inputfield for the Unityengine, the Input Values are saved in the List content
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
