--- Assets/Scripts/ActiveDeck.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b86fe8c738da44418f18f32b239f384
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
