--- Assets/Scripts/GMSkript.cs
@@ -1,27 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class GMSkript : MonoBehaviour
-{
-    [SerializeField] float delay = 1.0f;
-    [SerializeField] GameObject playerDeck;
-    [SerializeField] GameObject computerDeck;
-    float timer = 0.0f;
-    bool hasClicked = false;
-    bool playerWon = false;
-    bool isDraw = false;
-    List<GameObject> activeCards;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
