+++ Assets/Scripts/GameMasterSkript.cs
@@ -2,9 +2,16 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class MainDeckScript : MonoBehaviour
+public class GMSkript : MonoBehaviour
 {
-
+    [SerializeField] float delay = 1.0f;
+    [SerializeField] GameObject playerDeck;
+    [SerializeField] GameObject computerDeck;
+    float timer = 0.0f;
+    bool hasClicked = false;
+    bool playerWon = false;
+    bool isDraw = false;
+    List<GameObject> activeCards;
 
     [SerializeField] public List<GameObject> cardPile;
     public List<GameObject> playerPile;
@@ -37,20 +44,17 @@ public class MainDeckScript : MonoBehaviour
 
     }
 
-
+    // Start is called before the first frame update
     void Start()
     {
         // Shuffles and separates the cardPile
         Shuffle(cardPile);
         Seperate(cardPile);
-
     }
-
 
     // Update is called once per frame
     void Update()
     {
-
+        
     }
 }
-
