--- Assets/Scripts/MainDeckScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 49ee4fc9113829746bbb2bb43f472a73
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
